Prepare a docker environment for development based on trac
The docker environment can be prepared and run using the tools
found in the docker directory.  The docker images prepared
by these tools store their data in external volumes, these
data volumes are named according to “projects” and used
accordingly by the compose tool when running the environment.

The environment currently only consists of a trac server
and support scripts to add repositories and keep them in
sync with trac.

The trac server is accessible through a HA-PROXY load-balancer
to which more services can be later added.

Close #14.