Remove CLSQL as a persistence backend
It seems a good time reconsidering using CLSQL as a persistence
backend. Using readably printed objects seems a much simpler
and as efficient approach.

Using CLSQL as a persistence backend introduces a lot of coding
overhead to work nicely with joined slots.  This certainly could be
delegated to a macro but that approach requires to build up a solid
knowledge of CLSQL internals around joined slots as well as taking a
lot of architectural decisions — like defining methods for
CLSQL:UPDATE-RECORDS-FROM-INSTANCE that also work on prerequisites of
a resource, thus modifying the usual semantic of the method, or
introducing CID specific methods to expose the fucntionality to the
user.

Using readably printed objects is a lightweight approach to
persistence. It however hinders our capacity to easily study
resources transversally to infrastructure stacks. Questions such as
“How many underlying resources are common in two infrastructure
stacks” or “Would destroying an infrastructure stack have consequences
for another infrastructure stack?” cannot be as easily answered
without a SQL persistence backend. These questions are however a
secondary target.