Delete Docker networks after deleting the project
Docker compose projects automatically create the required networks
according to the configuration in the docker compose file. These
networks are however not being deleted by the commands

  docker compose rm …

or

  docker compose down …

These networks need to be removed because they consume a subset of the
IP space and an accumulation of idle project specific networks
prevents new projects to run.